refactor: remove unnecessary grouping of Article/buttons
