+++ src/components/article/ArticleBanner.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import { ArticleButtons } from '@/components/article/buttons';
+import ArticleButtons from '@/components/article/ArticleButtons';
 import { ArticleBannerProps } from './type';
 
 export default function ArticleBanner({ slug, title, profileLink, image, username, date }: ArticleBannerProps) {
