--- src/components/article/buttons/index.ts
@@ -1,4 +1,0 @@
-import CommentButtonDelete from './CommentButtonDelete';
-import ArticleButtons from './ArticleButtons';
-
-export { CommentButtonDelete, ArticleButtons };
